refactor: modularize apts/skyfield_searches/jovian.py into a package
Refactored the monolithic `apts/skyfield_searches/jovian.py` into a modular package `apts/skyfield_searches/jovian/` to improve maintainability and reduce cyclomatic complexity.

Key changes:
- Split logic into `grs.py` (GRS transits), `moons.py` (Galilean moon events), `mutual.py` (mutual events), and `utils.py` (shared helpers).
- De-nested `state_func`, `mutual_state`, and extracted `is_inside_ellipsoid_projection` to reduce complexity scores below the threshold of 10.
- Maintained backward compatibility for the public API via re-exports in `apts/skyfield_searches/jovian/__init__.py`.
- Updated unit test mocks in `tests/unit/test_jovian_mutual_events.py` to reflect the new module structure.
- Verified system-wide correctness with a full 663-test suite pass.
